requirements.txt
# ── API ───────────────────────────────────────────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# ── Numerical / ML ────────────────────────────────────────────────────────────
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# ── Graph Neural Network ──────────────────────────────────────────────────────
torch>=2.0.0
torch_geometric>=2.4.0

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy>=2.0.0

# ── GCP (gracefully disabled when absent) ────────────────────────────────────
google-cloud-monitoring>=2.18.0
google-cloud-storage>=2.10.0